+++ lib/icalendar2/property/tzoffsetfrom.rb
@@ -1,0 +1,9 @@
+module Icalendar2
+  module Property
+    # See http://tools.ietf.org/html/rfc5545#page-72
+    class Tzoffsetfrom < Base
+      name "TZOFFSETFROM"
+      value :types => [:text]
+    end
+  end
+end