+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+0.1.7 2013-03-05
+* Add timezone component
+
+0.1.6 2013-02-27
+* Fix for trigger properties
+
+0.1.5 2013-02-22
+* Add alarm component
+
+0.1.4 2013-02-05
+* Fix for components with arrays
+
+0.1.3 2013-02-05
+* Add categories property spec
+
 0.1.2 2013-01-31
 * Add created property